backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymongo==4.12.0
python-dotenv==1.0.1
